# Reinvent wheel for encoding json