--- a/go.mod
+++ b/go.mod
@@ -1,9 +1,6 @@
 module github.com/botobag/artemis
 
 require (
-	github.com/json-iterator/go v1.1.5
-	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
-	github.com/modern-go/reflect2 v1.0.1 // indirect
 	github.com/onsi/ginkgo v1.6.0
 	github.com/onsi/gomega v1.4.2
 )