The results of our micro-benchmark shows a 45% improvement in execution time with 64% reduction on number of allocations.

### Before

```bash
$ go test -benchmem -bench=BenchmarkMarshalStarWarsFriendsQueryResultToJSON -run=^$ .
goos: darwin
goarch: amd64
pkg: github.com/botobag/artemis/graphql/executor
BenchmarkMarshalStarWarsFriendsQueryResultToJSONWithGo-8                  100000             12092 ns/op            1986 B/op         16 allocs/op
BenchmarkMarshalStarWarsFriendsQueryResultToJSONWithJsoniter-8            200000              7702 ns/op            1976 B/op         31 allocs/op
PASS
ok      github.com/botobag/artemis/graphql/executor     2.987s
```

### After

```bash
$ go test -benchmem -bench=BenchmarkMarshalStarWarsFriendsQueryResultToJSON -run=^$ .
goos: darwin
goarch: amd64
pkg: github.com/botobag/artemis/graphql/executor
BenchmarkMarshalStarWarsFriendsQueryResultToJSONWithGo-8                  200000             10679 ns/op            3745 B/op         18 allocs/op
BenchmarkMarshalStarWarsFriendsQueryResultToJSONWithJsonwriter-8          300000              4199 ns/op            2656 B/op         11 allocs/op
PASS
ok      github.com/botobag/artemis/graphql/executor     3.585s
```